Better Documentation for Daemon Code

+++ srt/daemon/daemon.py
@@ -1,3 +1,9 @@
+"""daemon.py
+
+Main Control and Orchestration Class for the Small Radio Telescope
+
+"""
+
 from time import sleep, time
 from datetime import timedelta, datetime
 from threading import Thread
@@ -23,11 +29,25 @@ from .utilities.functions import azel_within_range
 
 
 class SmallRadioTelescopeDaemon:
+    """
+    Controller Class for the Small Radio Telescope
+    """
+
     def __init__(self, config_directory):
+        """Initializer for the Small Radio Telescope Daemon
+
+        Parameters
+        ----------
+        config_directory : str
+            Path to the Folder Containing the Configuration for the SRT
+        """
+
+        # Validate and Parse YAML Settings
         self.config_directory = config_directory
         validate_yaml_schema(path=config_directory)
         config_dict = load_yaml(path=config_directory)
 
+        # Store Individual Settings In Object
         self.station = config_dict["STATION"]
         self.contact = config_dict["EMERGENCY_CONTACT"]
         self.az_limits = (
@@ -52,11 +72,15 @@ class SmallRadioTelescopeDaemon:
         self.radio_sample_frequency = config_dict["RADIO_SF"]
         self.radio_num_bins = config_dict["RADIO_NUM_BINS"]
         self.radio_integ_cycles = config_dict["RADIO_INTEG_CYCLES"]
+        self.radio_autostart = config_dict["RADIO_AUTOSTART"]
         self.beamwidth = config_dict["BEAMWIDTH"]
         self.temp_sys = config_dict["TSYS"]
         self.temp_cal = config_dict["TCAL"]
         self.save_dir = config_dict["SAVE_DIRECTORY"]
 
+        # Generate Default Calibration Values
+        # Values are Set Up so that Uncalibrated and Calibrated Spectra are the Same Values
+        # Unless there is a pre-exisiting calibration from a previous run
         self.cal_values = [1.0 for _ in range(self.radio_num_bins)]
         self.cal_power = 1.0 / (self.temp_sys + self.temp_cal)
         calibration_path = Path(self.config_directory, "calibration.json")
@@ -64,12 +88,14 @@ class SmallRadioTelescopeDaemon:
             with open(calibration_path, "r") as input_file:
                 try:
                     cal_data = json.load(input_file)
+                    # If Calibration is of a Different Size Than The Current FFT Size, Discard
                     if len(cal_data["cal_values"]) == self.radio_num_bins:
                         self.cal_values = cal_data["cal_values"]
                         self.cal_power = cal_data["cal_pwr"]
                 except KeyError:
                     pass
 
+        # Create Helper Object Which Tracks Celestial Objects
         self.ephemeris_tracker = EphemerisTracker(
             self.station["latitude"],
             self.station["longitude"],
@@ -78,6 +104,7 @@ class SmallRadioTelescopeDaemon:
         self.ephemeris_locations = self.ephemeris_tracker.get_all_azimuth_elevation()
         self.ephemeris_cmd_location = None
 
+        # Create Rotor Command Helper Object
         self.rotor = Rotor(
             self.motor_type, self.motor_port, self.az_limits, self.el_limits,
         )
@@ -88,22 +115,46 @@ class SmallRadioTelescopeDaemon:
             map(add, self.rotor_destination, self.rotor_offsets)
         )
 
-        self.radio_queue = Queue()
+        # Create Radio Processing Task (Wrapper for GNU Radio Script)
         self.radio_process_task = RadioProcessTask(
             num_bins=self.radio_num_bins, num_integrations=self.radio_integ_cycles
         )
+        self.radio_queue = Queue()
         self.radio_save_raw_task = None
         self.radio_save_spec_task = None
 
+        # Create Object for Keeping Track of What Commands Are Running or Have Failed
         self.current_queue_item = "None"
         self.command_queue = Queue()
         self.command_error_logs = []
 
     def log_message(self, message):
+        """Writes Contents to a Logging List and Prints
+
+        Parameters
+        ----------
+        message : str
+            Message to Log and Print
+
+        Returns
+        -------
+        None
+        """
         self.command_error_logs.append((time(), message))
         print(message)
 
     def n_point_scan(self, object_id):
+        """Runs an N-Point (25) Scan About an Object
+
+        Parameters
+        ----------
+        object_id : str
+            Name of the Object to Perform N-Point Scan About
+
+        Returns
+        -------
+        None
+        """
         self.ephemeris_cmd_location = None
         self.radio_queue.put(("soutrack", object_id))
         for scan in range(25):
@@ -122,6 +173,17 @@ class SmallRadioTelescopeDaemon:
         self.ephemeris_cmd_location = object_id
 
     def beam_switch(self, object_id):
+        """Swings Antenna Across Object
+        
+        Parameters
+        ----------
+        object_id : str
+            Name of the Object to Perform Beam-Switch About
+
+        Returns
+        -------
+
+        """
         self.ephemeris_cmd_location = None
         self.radio_queue.put(("soutrack", object_id))
         new_rotor_destination = self.ephemeris_locations[object_id]
@@ -139,6 +201,17 @@ class SmallRadioTelescopeDaemon:
         self.ephemeris_cmd_location = object_id
 
     def point_at_object(self, object_id):
+        """Points Antenna Directly at Object, and Sets Up Tracking to Follow it
+
+        Parameters
+        ----------
+        object_id : str
+            Name of Object to Point at and Track
+
+        Returns
+        -------
+        None
+        """
         self.rotor_offsets = (0.0, 0.0)
         self.radio_queue.put(("soutrack", object_id))
         new_rotor_cmd_location = self.ephemeris_locations[object_id]
@@ -153,6 +226,19 @@ class SmallRadioTelescopeDaemon:
             self.ephemeris_cmd_location = None
 
     def point_at_azel(self, az, el):
+        """Points Antenna at a Specific Azimuth and Elevation
+
+        Parameters
+        ----------
+        az : float
+            Azimuth, in degrees, to turn antenna towards
+        el : float
+            Elevation, in degrees, to point antenna upwards at
+
+        Returns
+        -------
+        None
+        """
         self.ephemeris_cmd_location = None
         self.rotor_offsets = (0.0, 0.0)
         self.radio_queue.put(("soutrack", f"azel_{az}_{el}"))
@@ -167,6 +253,19 @@ class SmallRadioTelescopeDaemon:
             self.log_message(f"Object at {new_rotor_cmd_location} Not in Motor Bounds")
 
     def point_at_offset(self, az_off, el_off):
+        """From the Current Object or Position Pointed At, Move to an Offset of That Location
+
+        Parameters
+        ----------
+        az_off : float
+            Number of Degrees in Azimuth Offset
+        el_off : float
+            Number of Degrees in Elevation Offset
+
+        Returns
+        -------
+        None
+        """
         new_rotor_offsets = (az_off, el_off)
         new_rotor_cmd_location = tuple(
             map(add, self.rotor_destination, new_rotor_offsets)
@@ -180,6 +279,12 @@ class SmallRadioTelescopeDaemon:
             self.log_message(f"Offset {new_rotor_offsets} Out of Bounds")
 
     def stow(self):
+        """Moves the Antenna Back to Its Stow Location
+
+        Returns
+        -------
+        None
+        """
         self.ephemeris_cmd_location = None
         self.radio_queue.put(("soutrack", "at_stow"))
         self.rotor_offsets = (0.0, 0.0)
@@ -189,6 +294,12 @@ class SmallRadioTelescopeDaemon:
             sleep(0.1)
 
     def calibrate(self):
+        """Runs Calibration Processing and Pushes New Values to Processing Script
+
+        Returns
+        -------
+        None
+        """
         radio_cal_task = RadioCalibrateTask(
             self.radio_num_bins, self.radio_integ_cycles, self.config_directory,
         )
@@ -204,6 +315,17 @@ class SmallRadioTelescopeDaemon:
         self.log_message("Calibration Done")
 
     def start_recording(self, name):
+        """Starts Recording Data
+
+        Parameters
+        ----------
+        name : str
+            Name of the File to Be Recorded
+
+        Returns
+        -------
+        None
+        """
         if self.radio_save_raw_task is None and self.radio_save_spec_task is None:
             if name is None or not name.endswith(".rad"):
                 self.radio_save_raw_task = RadioSaveRawTask(
@@ -222,30 +344,74 @@ class SmallRadioTelescopeDaemon:
             self.log_message("Cannot Start Recording - Already Recording")
 
     def stop_recording(self):
+        """Stops Any Current Recording, if Running
+
+        Returns
+        -------
+        None
+        """
         if self.radio_save_raw_task is not None:
             self.radio_save_raw_task.terminate()
             self.radio_save_raw_task = None
+        if self.radio_save_spec_task is not None:
+            self.radio_save_spec_task.terminate()
+            self.radio_save_spec_task = None
 
     def set_freq(self, frequency):
-        self.radio_center_frequency = frequency
-        self.radio_queue.put(("freq", self.radio_center_frequency))
+        """Set the Frequency of the Processing Script
+
+        Parameters
+        ----------
+        frequency : float
+            Center Frequency, in Hz, to Set SDR to
+
+        Returns
+        -------
+        None
+        """
+        self.radio_center_frequency = frequency  # Set Local Value
+        self.radio_queue.put(
+            ("freq", self.radio_center_frequency)
+        )  # Push Update to GNU Radio
 
     def set_samp_rate(self, samp_rate):
+        """Set the Sample Rate of the Processing Script
+
+        Note that this stops any currently running raw saving tasks
+
+        Parameters
+        ----------
+        samp_rate : float
+            Sample Rate for the SDR in Hz
+
+        Returns
+        -------
+        None
+        """
         if self.radio_save_raw_task is not None:
             self.radio_save_raw_task.terminate()
         self.radio_sample_frequency = samp_rate
         self.radio_queue.put(("samp_rate", self.radio_sample_frequency))
-        if self.radio_save_raw_task is not None:
-            self.radio_save_raw_task = RadioSaveRawTask(
-                self.radio_sample_frequency, self.save_dir
-            )
-            self.radio_save_raw_task.start()
 
     def quit(self):
+        """Stops the Daemon Process
+
+        Returns
+        -------
+        None
+        """
         keep_running = False
         self.radio_queue.put(("is_running", keep_running))
 
     def update_ephemeris_location(self):
+        """Periodically Updates Object Locations for Tracking Sky Objects
+
+        Is Operated as an Infinite Looping Thread Function
+
+        Returns
+        -------
+        None
+        """
         last_updated_time = None
         while True:
             if last_updated_time is None or time() - last_updated_time > 10:
@@ -274,6 +440,14 @@ class SmallRadioTelescopeDaemon:
             sleep(1)
 
     def update_rotor_status(self):
+        """Periodically Sets Rotor Azimuth and Elevation and Fetches New Antenna Position
+
+        Is Operated as an Infinite Looping Thread Function
+
+        Returns
+        -------
+        None
+        """
         while True:
             try:
                 current_rotor_cmd_location = self.rotor_cmd_location
@@ -307,6 +481,14 @@ class SmallRadioTelescopeDaemon:
                 self.log_message(str(e))
 
     def update_command_queue(self):
+        """Waits for New Commands Coming in Over ZMQ PUSH/PULL
+
+        Is Operated as an Infinite Looping Thread Function
+
+        Returns
+        -------
+        None
+        """
         context = zmq.Context()
         command_port = 5556
         command_socket = context.socket(zmq.PULL)
@@ -315,6 +497,14 @@ class SmallRadioTelescopeDaemon:
             self.command_queue.put(command_socket.recv_string())
 
     def update_status(self):
+        """Periodically Publishes Daemon Status for Dashboard (or any other subscriber)
+
+        Is Operated as an Infinite Looping Thread Function
+
+        Returns
+        -------
+        None
+        """
         context = zmq.Context()
         status_port = 5555
         status_socket = context.socket(zmq.PUB)
@@ -343,6 +533,14 @@ class SmallRadioTelescopeDaemon:
             sleep(0.5)
 
     def update_radio_settings(self):
+        """Coordinates Sending XMLRPC Commands to the GNU Radio Script
+
+        Is Operated as an Infinite Looping Thread Function
+
+        Returns
+        -------
+        None
+        """
         rpc_server = ServerProxy("http://localhost:5557/")
         while True:
             method, value = self.radio_queue.get()
@@ -351,6 +549,14 @@ class SmallRadioTelescopeDaemon:
             sleep(0.1)
 
     def srt_daemon_main(self):
+        """Starts and Processes Commands for the SRT
+
+        Returns
+        -------
+        None
+        """
+
+        # Create Infinite Looping Threads
         ephemeris_tracker_thread = Thread(
             target=self.update_ephemeris_location, daemon=True
         )
@@ -359,12 +565,15 @@ class SmallRadioTelescopeDaemon:
         status_thread = Thread(target=self.update_status, daemon=True)
         radio_thread = Thread(target=self.update_radio_settings, daemon=True)
 
-        try:
-            self.radio_process_task.start()
-        except RuntimeError as e:
-            self.log_message(str(e))
-        sleep(5)
+        # If the GNU Radio Script Should be Running, Start It
+        if self.radio_autostart:
+            try:
+                self.radio_process_task.start()
+            except RuntimeError as e:
+                self.log_message(str(e))
+            sleep(5)
 
+        # Send Settings to the GNU Radio Script
         radio_params = {
             "Frequency": ("freq", self.radio_center_frequency),
             "Sample Rate": ("samp_rate", self.radio_sample_frequency),
@@ -381,6 +590,7 @@ class SmallRadioTelescopeDaemon:
             self.log_message(f"Setting {name}")
             self.radio_queue.put(radio_params[name])
 
+        # Start Infinite Looping Update Threads
         ephemeris_tracker_thread.start()
         rotor_pointing_thread.start()
         command_queueing_thread.start()
@@ -391,6 +601,7 @@ class SmallRadioTelescopeDaemon:
 
         while keep_running:
             try:
+                # Await Command for the SRT
                 self.current_queue_item = "None"
                 command = self.command_queue.get()
                 self.log_message(f"Running Command '{command}'")
@@ -436,11 +647,11 @@ class SmallRadioTelescopeDaemon:
                     self.point_at_offset(
                         float(command_parts[1]), float(command_parts[2])
                     )
-                elif command_name.isnumeric():
+                elif command_name.isnumeric():  # If Command is a Number, Sleep that Long
                     sleep(float(command_name))
                 elif command_name == "wait":
                     sleep(float(command_parts[1]))
-                elif command_name.split(":")[0] == "lst":
+                elif command_name.split(":")[0] == "lst":  # Wait Until Next Time H:M:S
                     time_string = command_name.replace("LST:", "")
                     time_val = datetime.strptime(time_string, "%H:%M:%S")
                     while time_val < datetime.utcfromtimestamp(time()):
@@ -449,7 +660,7 @@ class SmallRadioTelescopeDaemon:
                         time_val - datetime.utcfromtimestamp(time())
                     ).total_seconds()
                     sleep(time_delta)
-                elif len(command_name.split(":")) == 5:
+                elif len(command_name.split(":")) == 5:  # Wait Until Y:D:H:M:S
                     time_val = datetime.strptime(command_name, "%Y:%j:%H:%M:%S")
                     time_delta = (
                         time_val - datetime.utcfromtimestamp(time())
@@ -465,9 +676,11 @@ class SmallRadioTelescopeDaemon:
             except ConnectionRefusedError as e:
                 self.log_message(str(e))
 
+        # On End, Return to Stow and End Recordings
         self.stow()
         self.stop_recording()
-        self.radio_process_task.terminate()
+        if self.radio_autostart:
+            self.radio_process_task.terminate()
 
 
 if __name__ == "__main__":
